[COMPLETE] Presentation Day. Final Review

+++ app/src/main/java/com/example/taboada/rtdbexample3/model/Message.kt
@@ -8,7 +8,7 @@ data class Message(
     val id: String = UUID.randomUUID().toString(),
     val content: String = "",
     val priority: String = "",
-    val timestamp: Int = 0,
+    val timestamp: Long = 0,
     val flag: String = "",
     val userId: String = "",
     val authorImage: Int = if (flag == "me") AppIcon.peach else AppIcon.daisy
